fixed serializers and BudgetModel to use integers instead of float

+++ api2/serializers.py
@@ -10,8 +10,8 @@ class BudgetSerializer(serializers.ModelSerializer):
     name = serializers.CharField(
         max_length=20, validators=[UniqueValidator(queryset=Budget.objects.all())]
     )
-    percentage = serializers.FloatField(max_value=100)
-    initial_balance = serializers.FloatField(required=False)
+    percentage = serializers.IntegerField(max_value=100)
+    initial_balance = serializers.IntegerField(required=False)
     balance = serializers.SerializerMethodField(read_only=True)
     user = serializers.PrimaryKeyRelatedField(many=False, queryset=User.objects.all())
